--- pkg/agent/handler/discoverymanagedapplication.go
@@ -1,64 +1,0 @@
-package handler
-
-import (
-	"context"
-
-	agentcache "github.com/Axway/agent-sdk/pkg/agent/cache"
-	apiv1 "github.com/Axway/agent-sdk/pkg/apic/apiserver/models/api/v1"
-	management "github.com/Axway/agent-sdk/pkg/apic/apiserver/models/management/v1"
-	"github.com/Axway/agent-sdk/pkg/watchmanager/proto"
-)
-
-type discoveryManagedApplication struct {
-	marketplaceHandler
-	cache agentcache.Manager
-}
-
-// NewDiscoveryManagedApplicationHandler creates a Handler for ManagedApplications for discovery agent cache building
-func NewDiscoveryManagedApplicationHandler(cache agentcache.Manager) Handler {
-	return &discoveryManagedApplication{
-		cache: cache,
-	}
-}
-
-func (h *discoveryManagedApplication) ShouldHandle(ctx context.Context, event *proto.Event) bool {
-	cachedApp := h.cache.GetManagedApplication(event.Payload.Metadata.Id)
-	if cachedApp != nil {
-		return false
-	}
-
-	return true
-}
-
-// HandleCache builds the ManagedApplication cache during discoveryCache's bulk rebuild.
-func (h *discoveryManagedApplication) HandleCache(resource *apiv1.ResourceInstance) error {
-	h.cache.AddManagedApplication(resource)
-	return nil
-}
-
-// Handle processes events triggered for ManagedApplications during discovery cache building
-func (h *discoveryManagedApplication) Handle(ctx context.Context, _ *proto.EventMeta, resource *apiv1.ResourceInstance) error {
-	action := GetActionFromContext(ctx)
-	if action == proto.Event_DELETED {
-		return h.cache.DeleteManagedApplication(resource.Metadata.ID)
-	}
-
-	app := &management.ManagedApplication{}
-	err := app.FromInstance(resource)
-	if err != nil {
-		return err
-	}
-
-	ok := isStatusFound(app.Status)
-	if !ok {
-		return nil
-	}
-
-	if h.shouldProcessForAgent(app.Status, app.Metadata.State) {
-		cachedApp := h.cache.GetManagedApplication(resource.Metadata.ID)
-		if cachedApp == nil {
-			h.cache.AddManagedApplication(resource)
-		}
-	}
-	return nil
-}
