--- pkg/agent/handler/tracemanagedapplication.go
@@ -1,64 +1,0 @@
-package handler
-
-import (
-	"context"
-
-	agentcache "github.com/Axway/agent-sdk/pkg/agent/cache"
-	apiv1 "github.com/Axway/agent-sdk/pkg/apic/apiserver/models/api/v1"
-	management "github.com/Axway/agent-sdk/pkg/apic/apiserver/models/management/v1"
-	"github.com/Axway/agent-sdk/pkg/watchmanager/proto"
-)
-
-type traceManagedApplication struct {
-	marketplaceHandler
-	cache agentcache.Manager
-}
-
-// NewTraceManagedApplicationHandler creates a Handler for Access Requests for trace agent
-func NewTraceManagedApplicationHandler(cache agentcache.Manager) Handler {
-	return &traceManagedApplication{
-		cache: cache,
-	}
-}
-
-func (h *traceManagedApplication) ShouldHandle(ctx context.Context, event *proto.Event) bool {
-	cachedApp := h.cache.GetManagedApplication(event.Payload.Metadata.Id)
-	if cachedApp != nil {
-		return false
-	}
-
-	return true
-}
-
-// HandleCache builds the ManagedApplication cache during discoveryCache's bulk rebuild.
-func (h *traceManagedApplication) HandleCache(resource *apiv1.ResourceInstance) error {
-	h.cache.AddManagedApplication(resource)
-	return nil
-}
-
-// Handle processes grpc events triggered for ManagedApplications for trace agent
-func (h *traceManagedApplication) Handle(ctx context.Context, _ *proto.EventMeta, resource *apiv1.ResourceInstance) error {
-	action := GetActionFromContext(ctx)
-	if action == proto.Event_DELETED {
-		return h.cache.DeleteManagedApplication(resource.Metadata.ID)
-	}
-
-	app := &management.ManagedApplication{}
-	err := app.FromInstance(resource)
-	if err != nil {
-		return err
-	}
-
-	ok := isStatusFound(app.Status)
-	if !ok {
-		return nil
-	}
-
-	if h.shouldProcessForAgent(app.Status, app.Metadata.State) {
-		cachedApp := h.cache.GetManagedApplication(resource.Metadata.ID)
-		if cachedApp == nil {
-			h.cache.AddManagedApplication(resource)
-		}
-	}
-	return nil
-}
